Move shared key auth to its own file
